+++ ext/tree-sitter/tree-sitter.h
@@ -6,9 +6,12 @@
 #endif
 
 #include "ruby.h"
+#include "runtime/document.h"
 #include "ruby/encoding.h"
 
-#include "runtime/document.h"
+#include "document.h"
+#include "node.h"
+#include "point.h"
 
 #define CSTR2SYM(s) (ID2SYM(rb_intern((s))))
